+++ cypress/integration/layers/thematiclayer.cy.js
@@ -3,6 +3,7 @@ import {
     DRILL_UP,
     DRILL_DOWN,
     VIEW_PROFILE,
+    SHOW_LONG_LAT,
     expectContextMenuOptions,
 } from '../../elements/map_context_menu.js'
 import { ThematicLayer } from '../../elements/thematic_layer.js'
@@ -29,25 +30,17 @@ context('Thematic Layers', () => {
 
     const Layer = new ThematicLayer()
 
-    it('shows error in layer edit modal if no indicator group selected', () => {
+    it('shows error in layer edit modal if no data selected', () => {
         Layer.openDialog('Thematic').addToMap()
 
         Layer.validateDialogClosed(false)
 
-        cy.contains('Indicator group is required').should('be.visible')
-    })
-
-    it('shows error in layer edit modal if no indicator selected', () => {
-        Layer.openDialog('Thematic')
-            .selectIndicatorGroup(HIV_INDICATOR_GROUP)
-            .addToMap()
-
-        Layer.validateDialogClosed(false)
-        cy.contains('Indicator is required').should('be.visible')
+        cy.contains('Data is required').should('be.visible')
     })
 
     it('shows error in layer edit modal if no period selected', () => {
         Layer.openDialog('Thematic')
+            .selectItemType('Indicators')
             .selectIndicatorGroup(HIV_INDICATOR_GROUP)
             .selectIndicator(HIV_INDICATOR_NAME)
             .selectTab('Period')
@@ -98,6 +91,7 @@ context('Thematic Layers', () => {
 
         // Choropleth
         Layer.openDialog('Thematic')
+            .selectItemType('Indicators')
             .selectIndicatorGroup(HIV_INDICATOR_GROUP)
             .selectIndicator(HIV_INDICATOR_NAME)
             .selectTab('Period')
@@ -157,6 +151,7 @@ context('Thematic Layers', () => {
 
     it('adds a thematic layer for OU Bombali', () => {
         Layer.openDialog('Thematic')
+            .selectItemType('Indicators')
             .selectIndicatorGroup(HIV_INDICATOR_GROUP)
             .selectIndicator(HIV_INDICATOR_NAME)
             .selectTab('Period')
@@ -174,7 +169,7 @@ context('Thematic Layers', () => {
 
     it('adds user sub-units and a Chiefdom OU', () => {
         Layer.openDialog('Thematic')
-            .selectItemType('Data element')
+            .selectItemType('Data elements')
             .selectDataElementGroup(ANC_DATAELEMENT_GROUP)
             .selectDataElement(ANC_DATAELEMENT_NAME)
             .selectTab('Org Units')
@@ -200,6 +195,7 @@ context('Thematic Layers', () => {
 
     it('adds a thematic layer with start and end date', () => {
         Layer.openDialog('Thematic')
+            .selectItemType('Indicators')
             .selectIndicatorGroup(HIV_INDICATOR_GROUP)
             .selectIndicator(HIV_INDICATOR_NAME)
             .selectTab('Period')
@@ -221,6 +217,7 @@ context('Thematic Layers', () => {
 
     it('opens a thematic layer popup with data and nodata', () => {
         Layer.openDialog('Thematic')
+            .selectItemType('Indicators')
             .selectIndicatorGroup('Stock')
             .selectIndicator('BCG Stock PHU')
             .selectTab('Period')
@@ -269,7 +266,7 @@ context('Thematic Layers', () => {
             parseFloat(text.replace('Value: ', ''))
 
         Layer.openDialog('Thematic')
-            .selectItemType('Data element')
+            .selectItemType('Data elements')
             .selectDataElementGroup(ANC_DATAELEMENT_GROUP)
             .selectDataElement(ANC_DATAELEMENT_NAME)
             .selectTab('Period')
@@ -365,12 +362,12 @@ context('Thematic Layers', () => {
     it('available rendering strategies depend on selected periods', () => {
         Layer.openDialog('Thematic').selectTab('Period').removeAllPeriods()
         cy.get('input[value="SINGLE"]').should('not.be.disabled')
-        cy.get('input[value="TIMELINE"]').should('be.disabled')
+        cy.get('input[value="TIMELINE"]').should('not.be.disabled')
         cy.get('div').contains('Timeline').realHover()
         cy.contains('Select at least 2 periods or 1 multi-period.').should(
             'be.visible'
         )
-        cy.get('input[value="SPLIT_BY_PERIOD"]').should('be.disabled')
+        cy.get('input[value="SPLIT_BY_PERIOD"]').should('not.be.disabled')
         cy.get('div').contains('Split').realHover()
         cy.contains('Select at least 2 periods or 1 multi-period.').should(
             'be.visible'
@@ -402,15 +399,16 @@ context('Thematic Layers', () => {
         })
         cy.get('input[value="SINGLE"]').should('not.be.disabled')
         cy.get('input[value="TIMELINE"]').should('not.be.disabled')
-        cy.get('input[value="SPLIT_BY_PERIOD"]').should('be.disabled')
+        cy.get('input[value="SPLIT_BY_PERIOD"]').should('not.be.disabled')
         cy.get('div').contains('Split').realHover()
         cy.contains(
             'Only up to a total of 12 periods (including those in multi-periods) can be selected.'
         ).should('be.visible')
     })
 
-    it('adds a thematic layer with timeline period', () => {
+    it('adds two thematic layer with timeline period', () => {
         Layer.openDialog('Thematic')
+            .selectItemType('Indicators')
             .selectIndicatorGroup(ANC_INDICATOR_GROUP)
             .selectIndicator(ANC_INDICATOR_NAME)
             .selectTab('Period')
@@ -427,6 +425,10 @@ context('Thematic Layers', () => {
             })
 
         cy.get('[type="radio"]').should('have.length', 3)
+        cy.get('input[value="SINGLE"]').should('not.be.disabled')
+        cy.get('input[value="TIMELINE"]').should('not.be.disabled')
+        cy.get('input[value="SPLIT_BY_PERIOD"]').should('not.be.disabled')
+        cy.get('input[value="SINGLE"]').should('be.checked')
         cy.get('[type="radio"]').check('TIMELINE')
 
         cy.getByDataTest('dhis2-uicore-modalactions')
@@ -434,6 +436,62 @@ context('Thematic Layers', () => {
             .click()
 
         Layer.validateDialogClosed(true)
+
+        Layer.validateCardTitle(ANC_INDICATOR_NAME)
+
+        // check for 1 maps
+        getMaps().should('have.length', 1)
+        // check for 1 layer
+        cy.getByDataTest('sortable-layers-list')
+            .children()
+            .should('have.length', 1)
+
+        Layer.openDialog('Thematic')
+            .selectItemType('Indicators')
+            .selectIndicatorGroup(HIV_INDICATOR_GROUP)
+            .selectIndicator(HIV_INDICATOR_NAME)
+            .selectTab('Org Units')
+            .selectOu('Sierra Leone')
+            .selectTab('Period')
+
+        cy.get('input[value="SINGLE"]').should('not.be.disabled')
+        cy.get('input[value="TIMELINE"]').should('not.be.disabled')
+        cy.get('input[value="SPLIT_BY_PERIOD"]').should('be.disabled')
+        cy.get('div')
+            .contains(
+                'Show multiple maps in view, one for each period (max 12).'
+            )
+            .realHover()
+        cy.contains('Remove all other layers to add a split layer.').should(
+            'be.visible'
+        )
+
+        cy.get('input[value="TIMELINE"]').should('be.checked')
+        cy.getByDataTest('period-dimension-transfer-option-content')
+            .contains('Last 4 quarters')
+            .should('exist')
+        cy.getByDataTest('period-dimension-transfer-option-content')
+            .contains(CURRENT_YEAR - 1)
+            .should('exist')
+
+        Layer.selectTab('Style').selectBubbleMap()
+
+        cy.getByDataTest('color-scale').click()
+        cy.getByDataTest('color-scale').eq(3).click()
+
+        cy.getByDataTest('dhis2-uicore-modalactions')
+            .contains('Add layer')
+            .click()
+
+        // check for 1 maps
+        getMaps().should('have.length', 1)
+        // check for 2 layer
+        cy.getByDataTest('sortable-layers-list')
+            .children()
+            .should('have.length', 2)
+
+        // wait to make sure the maps are loaded
+        cy.wait(2000) // eslint-disable-line cypress/no-unnecessary-waiting
 
         // check that the first timeline period is shown in blue
         cy.get('.dhis2-map-period').should('be.visible')
@@ -468,23 +526,18 @@ context('Thematic Layers', () => {
             .and('have.css', 'fill-opacity', '1')
         cy.get('.play-icon').should('be.visible')
 
-        Layer.openDialog('Thematic').selectTab('Period')
-
-        cy.get('input[value="SINGLE"]').should('not.be.disabled')
-        cy.get('input[value="TIMELINE"]').should('be.disabled')
-        cy.get('div').contains('Timeline').realHover()
-        cy.contains('Remove the existing timeline to add a new one.').should(
-            'be.visible'
-        )
-        cy.get('input[value="SPLIT_BY_PERIOD"]').should('be.disabled')
-        cy.get('div').contains('Split').realHover()
-        cy.contains('Remove all other layers to add a split map view.').should(
-            'be.visible'
-        )
+        expectContextMenuOptions([
+            { name: DRILL_UP, disabled: true },
+            { name: DRILL_DOWN },
+            { name: VIEW_PROFILE },
+            { name: SHOW_LONG_LAT },
+        ])
     })
 
-    it('adds a thematic layer with split view period', () => {
+    it('adds two thematic layer with split view period', () => {
+        // add a first layer
         Layer.openDialog('Thematic')
+            .selectItemType('Indicators')
             .selectIndicatorGroup(ANC_INDICATOR_GROUP)
             .selectIndicator(ANC_INDICATOR_NAME)
             .selectTab('Org Units')
@@ -497,6 +550,10 @@ context('Thematic Layers', () => {
             })
 
         cy.get('[type="radio"]').should('have.length', 3)
+        cy.get('input[value="SINGLE"]').should('not.be.disabled')
+        cy.get('input[value="TIMELINE"]').should('not.be.disabled')
+        cy.get('input[value="SPLIT_BY_PERIOD"]').should('not.be.disabled')
+        cy.get('input[value="SINGLE"]').should('be.checked')
         cy.get('[type="radio"]').check('SPLIT_BY_PERIOD')
 
         cy.getByDataTest('dhis2-uicore-modalactions')
@@ -509,6 +566,58 @@ context('Thematic Layers', () => {
 
         // check for 3 maps
         getMaps().should('have.length', 3)
+        // check for 1 layer
+        cy.getByDataTest('sortable-layers-list')
+            .children()
+            .should('have.length', 1)
+
+        // add a second layer
+        Layer.openDialog('Thematic')
+            .selectItemType('Indicators')
+            .selectIndicatorGroup(HIV_INDICATOR_GROUP)
+            .selectIndicator(HIV_INDICATOR_NAME)
+            .selectTab('Org Units')
+            .selectOu('Sierra Leone')
+            .selectTab('Period')
+
+        cy.get('input[value="SINGLE"]').should('be.disabled')
+        cy.get('div')
+            .contains('Show periods as a combined layer. Data is aggregated.')
+            .realHover()
+        cy.contains('Remove all split layers to add a single layer.').should(
+            'be.visible'
+        )
+        cy.get('input[value="TIMELINE"]').should('be.disabled')
+        cy.get('div')
+            .contains('Show multiple periods as an interactive timeline.')
+            .realHover()
+        cy.contains('Remove all split layers to add a timeline layer.').should(
+            'be.visible'
+        )
+        cy.get('input[value="SPLIT_BY_PERIOD"]').should('not.be.disabled')
+
+        cy.get('input[value="SPLIT_BY_PERIOD"]').should('be.checked')
+        cy.getByDataTest('period-dimension-transfer-option-content')
+            .contains('Last 3 months')
+            .should('exist')
+
+        Layer.selectTab('Style').selectBubbleMap()
+
+        cy.getByDataTest('color-scale').click()
+        cy.getByDataTest('color-scale').eq(3).click()
+
+        cy.getByDataTest('dhis2-uicore-modalactions')
+            .contains('Add layer')
+            .click()
+
+        Layer.validateDialogClosed(true)
+
+        // check for 3 maps
+        getMaps().should('have.length', 3)
+        // check for 2 layers
+        cy.getByDataTest('sortable-layers-list')
+            .children()
+            .should('have.length', 2)
 
         // wait to make sure the maps are loaded
         cy.wait(2000) // eslint-disable-line cypress/no-unnecessary-waiting
@@ -532,56 +641,12 @@ context('Thematic Layers', () => {
             expression: '#{fbfJHSPpUQD}/2',
         }).then((response) => {
             expect(response.status).to.eq(201)
-
             const calculationUid = response.body.response.uid
 
-            // open thematic dialog
-            cy.getByDataTest('add-layer-button').click()
-            cy.getByDataTest('addlayeritem-thematic').click()
-
-            // choose "Calculation" in item type
-            cy.getByDataTest('thematic-layer-value-type-select').click()
-            cy.contains('Calculations').click()
-
-            // assert that the label on the Calculation select is "Calculation"
-            cy.getByDataTest('calculationselect-label').contains('Calculation')
-
-            // click to open the calculation select
-            cy.getByDataTest('calculationselect').click()
-
-            // check search box exists "Type to filter options"
-            cy.getByDataTest('dhis2-uicore-popper')
-                .find('input[type="text"]')
-                .should('have.attr', 'placeholder', 'Type to filter options')
-
-            // search for something that doesn't exist
-            cy.getByDataTest('dhis2-uicore-popper')
-                .find('input[type="text"]')
-                .type('foo')
-
-            cy.getByDataTest('dhis2-uicore-select-menu-menuwrapper')
-                .contains('No options found')
-                .should('be.visible')
-
-            // try search for something that exists
-            cy.getByDataTest('dhis2-uicore-popper')
-                .find('input[type="text"]')
-                .clear()
-
-            cy.getByDataTest('dhis2-uicore-popper')
-                .find('input[type="text"]')
-                .type(calculationName)
-
-            cy.getByDataTest('dhis2-uicore-select-menu-menuwrapper')
-                .contains(calculationName)
-                .should('be.visible')
-
-            // select the calculation and close dialog
-            cy.contains(calculationName).click()
-
-            cy.getByDataTest('dhis2-uicore-modalactions')
-                .contains('Add layer')
-                .click()
+            Layer.openDialog('Thematic')
+                .selectItemType('Calculations')
+                .selectDataItem(calculationName)
+            Layer.addToMap()
 
             // check the layer card title
             cy.getByDataTest('layercard')
@@ -601,7 +666,7 @@ context('Thematic Layers', () => {
 
     it('adds a thematic layer for data element', () => {
         Layer.openDialog('Thematic')
-            .selectItemType('Data element')
+            .selectItemType('Data elements')
             .selectDataElementGroup(ANC_DATAELEMENT_GROUP)
             .selectDataElement(ANC_DATAELEMENT_NAME)
             .addToMap()
@@ -611,7 +676,7 @@ context('Thematic Layers', () => {
         Layer.validateCardTitle(ANC_DATAELEMENT_NAME)
         cy.getByDataTest(`card-ANC1stvisit`)
             .findByDataTest('layerlegend-item')
-            .should('have.length', 5)
+            .should('have.length', 7)
 
         getMaps().should('have.length', 1)
     })
@@ -620,6 +685,7 @@ context('Thematic Layers', () => {
         cy.visit('/')
 
         Layer.openDialog('Thematic')
+            .selectItemType('Indicators')
             .selectIndicatorGroup(HIV_INDICATOR_GROUP)
             .selectIndicator(HIV_INDICATOR_NAME)
             .selectTab('Period')
@@ -666,6 +732,7 @@ context('Thematic Layers', () => {
 
     it('adds a thematic layer with a filter', () => {
         Layer.openDialog('Thematic')
+            .selectItemType('Indicators')
             .selectIndicatorGroup('ANC')
             .selectIndicator('ANC 1 Coverage')
             .selectTab('Filter')