+++ d2.config.js
@@ -47,7 +47,13 @@ const config = {
     },
 
     viteConfigExtensions: {
+        optimizeDeps: {
+            esbuildOptions: {
+                target: 'es2022',
+            },
+        },
         build: {
+            target: 'es2022',
             rollupOptions: {
                 output: {
                     manualChunks: {