+++ src/constants/dimension.js
@@ -1,4 +1,4 @@
-export const dimConf = {
+const baseDimConf = {
     data: {
         id: 'data',
         value: 'data',
@@ -52,11 +52,23 @@ export const dimConf = {
         objectName: 'di',
         itemType: 'PROGRAM_DATA_ELEMENT',
     },
+    programDataElementOption: {
+        value: 'programDataElementOption',
+        dimensionName: 'dx',
+        objectName: 'di',
+        itemType: 'PROGRAM_DATA_ELEMENT_OPTION',
+    },
     programAttribute: {
         value: 'programAttribute',
         dimensionName: 'dx',
         objectName: 'di',
         itemType: 'PROGRAM_ATTRIBUTE',
+    },
+    programAttributeOption: {
+        value: 'programAttribute',
+        dimensionName: 'dx',
+        objectName: 'di',
+        itemType: 'PROGRAM_ATTRIBUTE_OPTION',
     },
     programIndicator: {
         value: 'programIndicator',
@@ -98,3 +110,13 @@ export const dimConf = {
         objectName: 'value',
     },
 }
+
+const dimConf = Object.entries(baseDimConf).reduce((acc, [key, value]) => {
+    acc[key] = value
+    if (value.itemType) {
+        acc[value.itemType] = value
+    }
+    return acc
+}, {})
+
+export { dimConf }