+++ src/constants/earthEngineLayers/index.js
@@ -9,12 +9,14 @@ import humidityWeekly from './humidity_weekly_ERA5-Land.js'
 import landcover from './landcover_MCD12Q1.js'
 import legacyBuildings from './legacy/buildings_GOOGLE.js'
 import legacyNighttime from './legacy/nighttime_DMSP-OLS.js'
+import legacyPopulationAgeSex from './legacy/population_age_sex_Worldpop.js'
+import legacyPopulationTotal from './legacy/population_total_Worldpop.js'
 import legacyPopulationWorldPop from './legacy/population_WorldPop.js'
 import legacyPopulation100m from './legacy/population_WorldPop_100m.js'
 import legacyPrecipitation from './legacy/precipitation_pentad_CHIRPS.js'
 import legacyTemperature from './legacy/temperature_MOD11A2v061.js'
-import populationAgeSex from './population_age_sex_WorldPop.js'
-import populationTotal from './population_total_WorldPop.js'
+import populationAgeSexWorldpopGlobal2 from './population_age_sex_Worldpop-Global2.js'
+import populationTotalWorldpopGlobal2 from './population_total_Worldpop-Global2.js'
 import precipitationChirpsDaily from './precipitation_daily_CHIRPS.js'
 import precipitationEra5Daily from './precipitation_daily_ERA5-Land.js'
 import precipitationChirpsMonthly from './precipitation_monthly_CHIRPS.js'
@@ -29,8 +31,8 @@ import vegetationModisMonthly from './vegetation_monthly_MOD13Q1.js'
 import vegetationModisWeekly from './vegetation_weekly_MOD13Q1.js'
 
 const earthEngineLayersFns = [
-    populationTotal,
-    populationAgeSex,
+    populationTotalWorldpopGlobal2,
+    populationAgeSexWorldpopGlobal2,
     buildings,
     elevation,
     heatMonthly,
@@ -56,6 +58,8 @@ const earthEngineLayersFns = [
     legacyNighttime,
     legacyPopulation100m,
     legacyPopulationWorldPop,
+    legacyPopulationTotal,
+    legacyPopulationAgeSex,
     legacyPrecipitation,
     legacyTemperature,
 ]
@@ -63,14 +67,20 @@ const earthEngineLayersFns = [
 const getEarthEngineLayers = () => earthEngineLayersFns.map((l) => l())
 
 const earthEngineLayersDefaultFns = [
-    populationTotal,
-    populationAgeSex,
+    populationTotalWorldpopGlobal2,
+    populationAgeSexWorldpopGlobal2,
     buildings,
     elevation,
     precipitationEra5Monthly,
     temperatureMonthly,
     landcover,
 ]
+
+export const earthEngineLayersUpdates = {
+    [legacyPopulationTotal().layerId]: populationTotalWorldpopGlobal2().layerId, // v101.5.6
+    [legacyPopulationAgeSex().layerId]:
+        populationAgeSexWorldpopGlobal2().layerId, // v101.5.6
+}
 
 export const earthEngineLayersIds = () =>
     earthEngineLayersFns.map((l) => l().layerId)